+++ www/adjust.js
@@ -146,7 +146,7 @@ var Adjust = {
     },
 
     getSdkPrefix: function () {
-        return 'cordova4.29.1';
+        return 'cordova4.30.0';
     },
 
     addSessionCallbackParameter: function(key, value) {
@@ -220,6 +220,10 @@ var Adjust = {
            return;
         }
         callCordova('onPause');
+    },
+
+    checkForNewAttStatus: function() {
+        callCordova('checkForNewAttStatus');
     }
 };