+++ CHANGELOG.md
@@ -1,3 +1,23 @@
+### Version 4.30.0 (24th May 2022)
+
+#### Added
+- Added ability to mark your app as COPPA compliant. You can enable this setting by calling `setCoppaCompliantEnabled` method of `AdjustConfig` instance with boolean parameter `true`.
+- Added ability to mark your Android app as app for the kids in accordance to Google Play Families policies. You can enable this setting by calling `setPlayStoreKidsAppEnabled` method of `AdjustConfig` instance with boolean parameter `true`.
+- Added `checkForNewAttStatus` method to `Adjust` API to allow iOS apps to instruct to SDK to check if `att_status` might have changed in the meantime.
+- Added support for Unity ad revenue tracking.
+- Added support for Helium Chartboost ad revenue tracking.
+- Added support for Generic ad revenue tracking.
+
+#### Changed
+- Started using CocoaPods to pull native Adjust iOS SDK dependency.
+- Started using Maven to pull native Adjust Android SDK dependency.
+
+#### Native SDKs
+- [iOS@v4.30.0][ios_sdk_v4.30.0]
+- [Android@v4.30.1][android_sdk_v4.30.1]
+
+---
+
 ### Version 4.29.1 (23rd September 2021)
 #### Added
 - Added support for `Admost` ad revenue source.
@@ -563,6 +583,7 @@
 [ios_sdk_v4.28.0]: https://github.com/adjust/ios_sdk/tree/v4.28.0
 [ios_sdk_v4.29.4]: https://github.com/adjust/ios_sdk/tree/v4.29.4
 [ios_sdk_v4.29.6]: https://github.com/adjust/ios_sdk/tree/v4.29.6
+[ios_sdk_v4.30.0]: https://github.com/adjust/ios_sdk/tree/v4.30.0
 
 [android_sdk_v3.5.0]: https://github.com/adjust/android_sdk/tree/v3.5.0
 [android_sdk_v4.1.0]: https://github.com/adjust/android_sdk/tree/v4.1.0
@@ -591,3 +612,5 @@
 [android_sdk_v4.27.0]: https://github.com/adjust/android_sdk/tree/v4.27.0
 [android_sdk_v4.28.4]: https://github.com/adjust/android_sdk/tree/v4.28.4
 [android_sdk_v4.28.5]: https://github.com/adjust/android_sdk/tree/v4.28.5
+[android_sdk_v4.30.1]: https://github.com/adjust/android_sdk/tree/v4.30.1
+